refactor: extract error enums to dedicated error.rs files
Move inline Error enums from lint/mod.rs, migrate/mod.rs, and
workspace_init/mod.rs into dedicated error.rs sub-modules. Each parent
module re-exports via `pub use error::Error` so external callers are
unaffected. Keeps error definitions focused and mod.rs files leaner.
