+++ crates/libaipm/src/lint/mod.rs
@@ -7,9 +7,12 @@
 
 pub mod config;
 pub mod diagnostic;
+pub mod error;
 pub mod reporter;
 pub mod rule;
 pub mod rules;
+
+pub use error::Error;
 
 use std::path::PathBuf;
 
@@ -187,36 +190,6 @@ pub struct Outcome {
     pub warning_count: usize,
     /// Source types that were scanned.
     pub sources_scanned: Vec<String>,
-}
-
-/// Errors that can occur during linting.
-#[derive(Debug, thiserror::Error)]
-pub enum Error {
-    /// I/O error during filesystem access.
-    #[error("I/O error: {0}")]
-    Io(#[from] std::io::Error),
-
-    /// JSON parse error.
-    #[error("JSON parse error in {path}: {reason}")]
-    JsonParse {
-        /// Path to the file that failed to parse.
-        path: PathBuf,
-        /// Reason for the parse failure.
-        reason: String,
-    },
-
-    /// Frontmatter parse error.
-    #[error("frontmatter parse error in {path}: {reason}")]
-    FrontmatterParse {
-        /// Path to the file that failed to parse.
-        path: PathBuf,
-        /// Reason for the parse failure.
-        reason: String,
-    },
-
-    /// Discovery failed during recursive directory walking.
-    #[error(transparent)]
-    DiscoveryFailed(#[from] crate::discovery::Error),
 }
 
 #[cfg(test)]
