+++ crates/libaipm/src/migrate/mod.rs
@@ -12,6 +12,7 @@ pub mod copilot_skill_detector;
 pub mod detector;
 pub mod dry_run;
 pub mod emitter;
+pub mod error;
 pub mod hook_detector;
 pub mod mcp_detector;
 pub mod output_style_detector;
@@ -24,6 +25,8 @@ use std::collections::HashSet;
 use std::path::{Path, PathBuf};
 
 use crate::fs::Fs;
+
+pub use error::Error;
 
 /// What kind of artifact was detected.
 #[derive(Debug, Clone, PartialEq, Eq, Hash)]
@@ -227,57 +230,6 @@ impl Outcome {
             })
             .collect()
     }
-}
-
-/// Errors specific to migration.
-#[derive(Debug, thiserror::Error)]
-pub enum Error {
-    /// The `.ai/` marketplace directory does not exist.
-    #[error("marketplace directory does not exist at {0} — run `aipm init --marketplace` first")]
-    MarketplaceNotFound(PathBuf),
-
-    /// The source directory does not exist.
-    #[error("source directory does not exist: {0}")]
-    SourceNotFound(PathBuf),
-
-    /// The source type is not supported.
-    #[error("unsupported source type '{0}' — supported sources: .claude, .github")]
-    UnsupportedSource(String),
-
-    /// Failed to parse marketplace.json.
-    #[error("failed to parse marketplace.json at {path}: {source}")]
-    MarketplaceJsonParse {
-        /// Path to the marketplace.json file.
-        path: PathBuf,
-        /// The underlying parse error.
-        source: serde_json::Error,
-    },
-
-    /// Failed to parse SKILL.md frontmatter.
-    #[error("failed to parse SKILL.md frontmatter in {path}: {reason}")]
-    FrontmatterParse {
-        /// Path to the SKILL.md file.
-        path: PathBuf,
-        /// Description of the parse failure.
-        reason: String,
-    },
-
-    /// Failed to parse a JSON configuration file.
-    #[error("failed to parse {path}: {reason}")]
-    ConfigParse {
-        /// Path to the configuration file.
-        path: PathBuf,
-        /// Description of the parse failure.
-        reason: String,
-    },
-
-    /// Discovery failed during recursive directory walking.
-    #[error("failed to discover source directories: {0}")]
-    DiscoveryFailed(#[from] crate::discovery::Error),
-
-    /// An I/O error occurred.
-    #[error(transparent)]
-    Io(#[from] std::io::Error),
 }
 
 /// Data needed to register a plugin in `marketplace.json`.
