+++ research/feature-list.json
@@ -376,7 +376,7 @@
       "Verify init.rs and cache.rs keep inline errors (1-2 variants)",
       "Run cargo build --workspace && cargo test --workspace && cargo clippy --workspace -- -D warnings && cargo fmt --check"
     ],
-    "passes": false
+    "passes": true
   },
   {
     "category": "refactor",
