+++ research/progress.txt
@@ -23,3 +23,4 @@
 [2026-04-13] Phase 3.2 COMPLETE — Consolidated lint rule helpers. Moved locate_json_key() to rules/mod.rs as pub(crate), deleted duplicates from hook_unknown_event.rs and hook_legacy_event.rs. Created simple_diag() in rules/mod.rs as pub(crate), updated 4 marketplace/plugin rules to delegate to it. Updated plugin_missing_manifest.rs inline Diagnostic to use simple_diag(). 7 new unit tests for shared helpers. 1797 tests pass. cargo build/test/clippy/fmt: PASS.
 [2026-04-13] Phase 3.3 COMPLETE — Extracted shared check_file() preamble helpers. Created read_skill_preamble(), read_agent_preamble(), read_hook_preamble() in rules/mod.rs. Updated 7 skill rules, 1 agent rule, 2 hook rules to use shared helpers. Removed direct `use super::scan;` imports from 10 rule files. 6 new unit tests. 1803 tests pass. cargo build/test/clippy/fmt: PASS.
 [2026-04-13] Phase 4.1 COMPLETE — Moved wizard types (PromptStep, PromptKind, PromptAnswer, styled_render_config) to crates/libaipm/src/wizard.rs behind #[cfg(feature = "wizard")] flag. Added wizard feature to libaipm Cargo.toml (dep:inquire optional). Both aipm and aipm-pack Cargo.toml now declare features = ["wizard"]. Both crate wizard.rs files replaced local type defs with pub use libaipm::wizard re-exports. Added Confirm arm to aipm-pack wizard.rs format_steps + wizard_tty.rs execute_prompts for exhaustive match. 4 unit tests in libaipm wizard.rs. 1807 tests pass. cargo build/test/clippy/fmt: PASS.
+[2026-04-13] Phase 4.2 COMPLETE — Extracted error enums to dedicated error.rs files for lint, migrate, and workspace_init modules. Created lint/error.rs (4 variants), migrate/error.rs (8 variants), workspace_init/error.rs (4 variants). Each parent mod.rs adds `pub mod error; pub use error::Error;`. Removed unused PathBuf import from workspace_init/mod.rs. init.rs and cache.rs verified as flat modules (no structural benefit from extraction). No .map_err patterns matched the #[from] replacement criteria. cargo build/test/clippy/fmt: PASS.
